everything working yay
