+++ studentside/backend/students.json
@@ -7,7 +7,11 @@
             "branch": "Civil Engineering",
             "degree": "BTech",
             "year": 3,
-            "applied": []
+            "applied": [
+                "AI-Powered Chatbot",
+                "Cybersecurity Threat Detection Using Machine Learning",
+                "Sustainable Concrete Using Recycled Materials"
+            ]
         },
         {
             "roll": "122201003",
@@ -16,7 +20,12 @@
             "branch": "Electrical Engineering",
             "degree": "BTech",
             "year": 3,
-            "applied": []
+            "applied": [
+                "AI-Powered Chatbot",
+                "IoT-Based Smart Home Automation",
+                "AI-Based Predictive Maintenance for Machines",
+                "Structural Health Monitoring Using IoT"
+            ]
         },
         {
             "roll": "142301001",
@@ -31,7 +40,8 @@
                 "Energy-Efficient Smart Grid System",
                 "Blockchain-Based Voting System",
                 "IoT-Based Smart Home Automation",
-                "Autonomous Vehicle Design"
+                "Autonomous Vehicle Design",
+                "Cybersecurity Threat Detection Using Machine Learning"
             ]
         }
     ]
