+++ studentside/src/app/app.routes.ts
@@ -1,8 +1,8 @@
+import { HomeComponent } from './home/home.component';
 import { InputComponent } from './input/input.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
 import { Routes } from '@angular/router';
 
 export const routes: Routes = [
     { path: '', component: InputComponent }, // Default route is the login page
-    { path: 'dashboard', component: DashboardComponent },
+    { path: 'dashboard', component: HomeComponent },
 ];
