--- studentside/src/app/dashboard/dashboard.component.html
@@ -1,79 +1,0 @@
-
-
-<div class="container-fluid mt-3">
-  <div class="card shadow-lg">
-    <div class="card-body">
-      <!-- Tabs Navigation -->
-      <ul class="nav nav-tabs">
-        <li class="nav-item">
-          <button class="nav-link" [class.active]="selectedTab === 'all'" (click)="selectTab('all')">All Projects</button>
-        </li>
-        <li class="nav-item">
-          <button class="nav-link" [class.active]="selectedTab === 'eligible'" (click)="selectTab('eligible')">Eligible Projects</button>
-        </li>
-        <li class="nav-item">
-          <button class="nav-link" [class.active]="selectedTab === 'profile'" (click)="selectTab('profile')">Profile</button>
-        </li>
-      </ul>
-
-      <div class="mt-4 ">
-        <div *ngIf="selectedTab === 'all' || selectedTab === 'eligible'">
-          <div class="table-responsive p-1">
-            <table class="table table-striped table-hover table-bordered align-middle rounded overflow-hidden">
-              <thead class="table rounded-lg">
-                <tr class="text-center">
-                  <th class="bg-primary text-white">Project Name</th>
-                  <th class="bg-primary text-white">Faculty</th>
-                  <th class="bg-primary text-white">Branch</th>
-                  <th class="bg-primary text-white">Year</th>
-                  <th class="bg-primary text-white">Action</th>
-                </tr>
-              </thead>
-              <tbody>
-                <ng-container *ngFor="let project of (selectedTab === 'all' ? allProjects : eligibleProjectDetails)">
-                  <tr (click)="toggleProject(project.name)" class="table-row">
-                    <td>{{ project.name }}</td>
-                    <td>{{ project.faculty }}</td>
-                    <td>{{ project.branch.join(', ') }}</td>
-                    <td>{{ project.year.join(', ') }}</td>
-                    <td>
-                      <button class="btn btn-sm"
-                        [ngClass]="applied.includes(project.name) ? 'btn-success' : 'btn-primary'"
-                        [disabled]="applied.includes(project.name)"
-                        (click)="applyForProject(project.name, $event)">
-                        {{ applied.includes(project.name) ? 'Applied' : 'Apply' }}
-                      </button>
-                    </td>
-                  </tr>
-                  <!-- Project Details Row -->
-                  <tr *ngIf="selectedProject === project.name">
-                    <td colspan="5" class="bg-light p-3">
-                      <div class="border rounded p-3 shadow-sm">
-                        <h5 class="text-primary">{{ project.name }}</h5>
-                        <p>{{ project.description }}</p>
-                      </div>
-                    </td>
-                  </tr>
-                </ng-container>
-              </tbody>
-            </table>
-          </div>
-        </div>
-
-        <div *ngIf="selectedTab === 'profile'" class="mt-3">
-          <div class="card shadow-sm">
-            <div class="card-body">
-              <h4 class="text-center text-primary">Student Profile</h4>
-              <hr>
-              <p><strong>Name:</strong> {{ name }}</p>
-              <p><strong>Roll Number:</strong> {{ roll }}</p>
-              <p><strong>Branch:</strong> {{ branch }}</p>
-              <p><strong>Year:</strong> {{ year }}</p>
-              <p><strong>Degree:</strong> {{ degree }}</p>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
