+++ studentside/src/app/input/input.component.html
@@ -1,6 +1,3 @@
-
-
-
 <div class="container d-flex justify-content-center align-items-center vh-100">
   <div class="card shadow-lg p-3" style="width: 400px">
     <div class="card-body">
