+++ studentside/src/app/project.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ProjectdataService } from './projectdata.service';
+import { projectService } from './project.service';
 
-describe('ProjectdataService', () => {
-  let service: ProjectdataService;
+describe('projectService', () => {
+  let service: projectService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(ProjectdataService);
+    service = TestBed.inject(projectService);
   });
 
   it('should be created', () => {
