--- studentside/src/app/sidepanel/sidepanel.component.html
@@ -1,22 +1,0 @@
-
-
-    <div class="container-fluid mt-3">
-    <div class="card shadow-lg rounded-lg">
-      <div class="card-header bg-primary text-white text-center fw-bold fs-5">
-        Application Tracker
-      </div>
-      <div class="card-body">
-        <ul class="list-group">
-          <li 
-            *ngFor="let project of applied" 
-            class="list-group-item d-flex justify-content-between align-items-center">
-            <span>{{ project }}</span>
-            <!-- <span class="badge bg-success">Applied</span> -->
-            <button class="btn btn-danger btn-sm" (click)="removeProject(project)">
-              Remove
-            </button>
-          </li>
-        </ul>
-      </div>
-    </div>
-    </div>
